add isExternal and separate out listings e2e tests
