+++ backend/core/test/listings/listings.e2e-spec.ts
@@ -453,158 +453,6 @@ describe("Listings", () => {
     })
   })
 
-  // REMOVE_WHEN_EXTERNAL_NOT_NEEDED
-  describe("/listings/combined", () => {
-    it("should return all listings", async () => {
-      const res = await supertest(app.getHttpServer()).get("/listings/combined").expect(200)
-      expect(res.body.items.map((listing) => listing.id).length).toBeGreaterThan(0)
-    })
-
-    it("should return the first page of paginated listings", async () => {
-      // Make the limit 1 less than the full number of listings, so that the first page contains all
-      // but the last listing.
-      const page = "1"
-      // This is the number of listings in ../../src/seed.ts minus 1
-      const limit = 15
-      const params = "/?page=" + page + "&limit=" + limit.toString()
-      const res = await supertest(app.getHttpServer())
-        .get("/listings/combined" + params)
-        .expect(200)
-      expect(res.body.items.length).toEqual(limit)
-    })
-
-    it("should return the last page of paginated listings", async () => {
-      // Make the limit 1 less than the full number of listings, so that the second page contains
-      // only one listing.
-      // query to get max number of listings
-      let queryParams = {
-        limit: 1,
-        page: 1,
-        view: "base",
-      }
-      let query = qs.stringify(queryParams)
-      let res = await supertest(app.getHttpServer()).get(`/listings/combined?${query}`).expect(200)
-      const totalItems = res.body.meta.totalItems
-
-      queryParams = {
-        limit: totalItems - 1,
-        page: 2,
-        view: "base",
-      }
-      query = qs.stringify(queryParams)
-      res = await supertest(app.getHttpServer()).get(`/listings/combined?${query}`).expect(200)
-      expect(res.body.items.length).toEqual(1)
-    })
-
-    it("should return listings with matching zipcodes", async () => {
-      const queryParams = {
-        limit: "all",
-        filter: [
-          {
-            $comparison: "IN",
-            zipcode: "94621,94404",
-          },
-        ],
-        view: "base",
-      }
-      const query = qs.stringify(queryParams)
-      await supertest(app.getHttpServer()).get(`/listings/combined?${query}`).expect(200)
-    })
-
-    it("should have listings associated with the Bay Area", async () => {
-      const jurisdictions = await jurisdictionsRepository.find()
-      const bayArea = jurisdictions.find((jurisdiction) => jurisdiction.name === "Bay Area")
-      const queryParamsOnlyBayArea = {
-        limit: "all",
-        view: "base",
-        filter: [
-          {
-            $comparison: "=",
-            jurisdiction: bayArea.id,
-          },
-        ],
-      }
-      const resBayArea = await supertest(app.getHttpServer())
-        .get(`/listings/combined?${qs.stringify(queryParamsOnlyBayArea)}`)
-        .expect(200)
-      expect(resBayArea.body.items.length).not.toBe(0)
-    })
-
-    it("defaults to sorting listings by applicationDueDate", async () => {
-      const res = await supertest(app.getHttpServer())
-        .get(`/listings/combined?limit=all`)
-        .expect(200)
-      const listings = res.body.items
-
-      // The Coliseum seed has the soonest applicationDueDate (1 day in the future)
-      expect(listings[0].name).toBe("[doorway] Test: Coliseum")
-
-      // Triton and "Default, No Preferences" share the next-soonest applicationDueDate
-      const secondListing = listings[1]
-      const thirdListing = listings[2]
-      expect(thirdListing.name).toBe("[doorway] Test: Default, No Preferences")
-
-      const secondListingAppDueDate = new Date(secondListing.applicationDueDate)
-      const thirdListingAppDueDate = new Date(thirdListing.applicationDueDate)
-      expect(secondListingAppDueDate.getDate()).toEqual(thirdListingAppDueDate.getDate())
-
-      // Verify that listings with null applicationDueDate's appear at the end.
-      const lastListing = listings[listings.length - 1]
-      expect(lastListing.applicationDueDate).toBeNull()
-    })
-
-    it("sorts listings by most recently updated when that orderBy param is set", async () => {
-      const res = await supertest(app.getHttpServer())
-        .get(`/listings/combined?orderBy[0]=mostRecentlyUpdated&orderDir[0]=DESC&limit=all`)
-        .expect(200)
-      for (let i = 0; i < res.body.items.length - 1; ++i) {
-        const currentUpdatedAt = new Date(res.body.items[i].updatedAt)
-        const nextUpdatedAt = new Date(res.body.items[i + 1].updatedAt)
-
-        // Verify that each listing's updatedAt timestamp is more recent than the next listing's.
-        expect(currentUpdatedAt.getTime()).toBeGreaterThan(nextUpdatedAt.getTime())
-      }
-    })
-
-    it("fails if orderBy param doesn't conform to one of the enum values", async () => {
-      await supertest(app.getHttpServer())
-        .get(`/listings/combined?orderBy=notAValidOrderByParam`)
-        .expect(400)
-    })
-
-    it("sorts results within a page, and across sequential pages", async () => {
-      // Get the first page of 5 results.
-      const firstPage = await supertest(app.getHttpServer())
-        .get(`/listings/combined?orderBy[0]=mostRecentlyUpdated&orderDir[0]=DESC&limit=5&page=1`)
-        .expect(200)
-
-      // Verify that listings on the first page are ordered from most to least recently updated.
-      for (let i = 0; i < 4; ++i) {
-        const currentUpdatedAt = new Date(firstPage.body.items[i].updatedAt)
-        const nextUpdatedAt = new Date(firstPage.body.items[i + 1].updatedAt)
-
-        // Verify that each listing's updatedAt timestamp is more recent than the next listing's.
-        expect(currentUpdatedAt.getTime()).toBeGreaterThan(nextUpdatedAt.getTime())
-      }
-
-      const lastListingOnFirstPageUpdateTimestamp = new Date(firstPage.body.items[4].updatedAt)
-
-      // Get the second page of 5 results
-      const secondPage = await supertest(app.getHttpServer())
-        .get(`/listings/combined?orderBy[0]=mostRecentlyUpdated&orderDir[0]=DESC&limit=5&page=2`)
-        .expect(200)
-
-      // Verify that each of the listings on the second page was less recently updated than the last
-      // first-page listing.
-      for (const secondPageListing of secondPage.body.items) {
-        const secondPageListingUpdateTimestamp = new Date(secondPageListing.updatedAt)
-        expect(lastListingOnFirstPageUpdateTimestamp.getTime()).toBeGreaterThan(
-          secondPageListingUpdateTimestamp.getTime()
-        )
-      }
-    })
-  })
-
   afterEach(() => {
     jest.clearAllMocks()
   })
